+++ include/mdpt.h
@@ -13,7 +13,6 @@ void kinetic_explicit_acc(T_prec& temp_out, energy_prec& eksum_out,
 void kinetic_explicit_cu(T_prec& temp_out, energy_prec& eksum_out,
                          energy_prec (&ekin_out)[3][3], const vel_prec* vx,
                          const vel_prec* vy, const vel_prec* vz);
-void kinetic_leapfrog(T_prec& temp);
 
 
 void temper(time_prec dt, T_prec& temp, bool save);
