+++ src/mdpq.cpp
@@ -311,79 +311,4 @@ void vel_data(rc_op op)
       wait_for(g::q0);
    }
 }
-
-
-//====================================================================//
-
-
-void swap_velocity(vel_prec* vxnew, vel_prec* vynew, vel_prec* vznew,
-                   vel_prec* vxold, vel_prec* vyold, vel_prec* vzold)
-{
-   swap_velocity_acc(vxnew, vynew, vznew, vxold, vyold, vzold);
-}
-
-
-void propagate_pos_lp(time_prec dt, pos_prec* x_lp, pos_prec* y_lp,
-                      pos_prec* z_lp, const vel_prec* vx_lp,
-                      const vel_prec* vy_lp, const vel_prec* vz_lp,
-                      const pos_prec* xold_lp, const pos_prec* yold_lp,
-                      const pos_prec* zold_lp, double scale)
-{
-
-   propagate_pos_lp_acc(dt, x_lp, y_lp, z_lp, vx_lp, vy_lp, vz_lp, xold_lp,
-                        yold_lp, zold_lp, scale);
-}
-
-
-void propagate_pos_lp2(time_prec dt, const pos_prec* x_lp, const pos_prec* y_lp,
-                       const pos_prec* z_lp, pos_prec* xold_lp,
-                       pos_prec* yold_lp, pos_prec* zold_lp, double scale)
-{
-   propagate_pos_lp2_acc(dt, x_lp, y_lp, z_lp, xold_lp, yold_lp, zold_lp,
-                         scale);
-}
-
-
-void propagate_pos_lf(time_prec dt, pos_prec* qx, pos_prec* qy, pos_prec* qz,
-                      const pos_prec* qxold, const pos_prec* qyold,
-                      const pos_prec* qzold, const vel_prec* vlx,
-                      const vel_prec* vly, const vel_prec* vlz)
-{
-   propagate_pos_lf_acc(dt, qx, qy, qz, qxold, qyold, qzold, vlx, vly, vlz);
-}
-
-
-void propagate_velocity_lp(vel_prec* vx_lp, vel_prec* vy_lp, vel_prec* vz_lp,
-                           const vel_prec* vxnew_lp, const vel_prec* vynew_lp,
-                           const vel_prec* vznew_lp, const vel_prec* vxold_lp,
-                           const vel_prec* vyold_lp, const vel_prec* vzold_lp,
-                           const double scale, energy_prec& eksum_new,
-                           energy_prec& eksum_old)
-{
-   propagate_velocity_lp_acc(vx_lp, vy_lp, vz_lp, vxnew_lp, vynew_lp, vznew_lp,
-                             vxold_lp, vyold_lp, vzold_lp, scale, eksum_new,
-                             eksum_old);
-}
-
-
-void propagate_velocity_lp2(time_prec dt, vel_prec* vx_lp, vel_prec* vy_lp,
-                            vel_prec* vz_lp, const pos_prec* x_lp,
-                            const pos_prec* y_lp, const pos_prec* z_lp,
-                            const pos_prec* xold_lp, const pos_prec* yold_lp,
-                            const pos_prec* zold_lp)
-{
-   propagate_velocity_lp2_acc(dt, vx_lp, vy_lp, vz_lp, x_lp, y_lp, z_lp,
-                              xold_lp, yold_lp, zold_lp);
-}
-
-
-void propagate_velocity_lp3(vel_prec* vx_lp, vel_prec* vy_lp, vel_prec* vz_lp,
-                            const vel_prec* vxnew_lp, const vel_prec* vynew_lp,
-                            const vel_prec* vznew_lp, const vel_prec* vxold_lp,
-                            const vel_prec* vyold_lp, const vel_prec* vzold_lp,
-                            energy_prec& eksum_new)
-{
-   propagate_velocity_lp3_acc(vx_lp, vy_lp, vz_lp, vxnew_lp, vynew_lp, vznew_lp,
-                              vxold_lp, vyold_lp, vzold_lp, eksum_new);
-}
 }
